backend/requirements.txt
# Requerimientos para ETL OtakuLATAM
# Instalación: pip install -r requirements.txt

# Manipulación de datos
pandas>=1.5.0
numpy>=1.21.0

# Base de datos
mysql-connector-python>=8.0.0

# Archivos Parquet
pyarrow>=10.0.0

# HTTP requests para TRM
requests>=2.28.0

# API REST y servidor web
Flask>=2.3.0
Flask-CORS>=4.0.0  # Para permitir requests desde frontend
Werkzeug>=2.3.0

# Almacenamiento en nube (opcional)
boto3>=1.26.0  # AWS S3
google-cloud-storage>=2.7.0  # Google Cloud Storage
azure-storage-blob>=12.14.0  # Azure Blob Storage

# Utilidades
python-dotenv>=0.19.0  # Variables de entorno
pathlib2>=2.3.0  # Compatibilidad de rutas

# Logging y monitoreo
colorlog>=6.7.0  # Logs con colores

# Desarrollo y testing (opcional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0  # Formateo de código
flake8>=5.0.0  # Linting
